+++ deal/linter/_extractors/__init__.py
@@ -2,6 +2,7 @@
 from .asserts import get_asserts
 from .common import get_name
 from .contracts import get_contracts
+from .definitions import get_definitions
 from .exceptions import get_exceptions
 from .imports import get_imports
 from .markers import get_markers
@@ -13,6 +14,7 @@ from .value import get_value
 __all__ = [
     'get_asserts',
     'get_contracts',
+    'get_definitions',
     'get_exceptions',
     'get_imports',
     'get_markers',
