+++ tests/test_linter/test_func.py
@@ -54,38 +54,3 @@ def test_repr():
     funcs2 = Func.from_astroid(astroid.parse(TEXT))
     for func in (funcs1[0], funcs2[0]):
         assert repr(func) == 'Func(post, raises)'
-
-
-@pytest.mark.parametrize('source, names', [
-    ('import re', {'re'}),
-    ('import typing, types', {'typing', 'types'}),
-    ('import typing as types', {'types'}),
-
-    ('from typing import List', {'List'}),
-    ('from typing import List, Dict', {'List', 'Dict'}),
-
-    ('ab = 2', {'ab'}),
-    ('ab = cd = 23', {'ab', 'cd'}),
-])
-def test_extract_defs(source: str, names) -> None:
-    tree = ast.parse(source)
-    print(ast.dump(tree))
-    defs = Func._extract_defs_ast(tree)
-    assert set(defs) == names
-
-    module = ast.parse('hello')
-    for name, stmt in defs.items():
-        module.body[0] = stmt
-        print(name, '|>', ast.dump(module))
-        compile(module, filename='<ast>', mode='exec')
-
-    tree = astroid.parse(source)
-    print(tree.repr_tree())
-    defs = Func._extract_defs_astroid(tree)
-    assert set(defs) == names
-
-    module = ast.parse('hello')
-    for name, stmt in defs.items():
-        module.body[0] = stmt
-        print(name, '|>', ast.dump(module))
-        compile(module, filename='<ast>', mode='exec')
